Update README.md yalan2ny
Pull request and file integrity and adding updates and comments to fix recent issues. My problem is the failure to deploy tokens to endpoints because the token list is displayed in TonViewer. It is also deployed to the Ton Battery account but without access. TonViewer also does not provide any path to access. No funds have been deployed to the user account in TonKeeper wallet so far. It should be noted that our devices are Android operating system with Chrome browser.